fix: resolve all clippy and -Dwarnings errors for CI
- Fix unused imports/variables across test files
- Fix digit grouping in nanosecond literals
- Fix approximate float constants (use std::f64::consts)
- Fix needless borrows, unnecessary casts, redundant closures
- Add PathBuf cfg(linux) guard in io_uring_reader
- Restore ClusterNode import in cluster/router
- All checks pass: fmt, clippy -Dwarnings, cargo check -Dwarnings

+++ crates/core/src/table.rs
@@ -1533,7 +1533,7 @@ mod tests {
         let n = 1000;
         let base_ts: i64 = 1_704_067_200_000_000_000;
         let timestamps: Vec<i64> = (0..n).map(|i| base_ts + i as i64 * 1_000_000).collect();
-        let symbols: Vec<i32> = (0..n).map(|i| (i % 100) as i32).collect();
+        let symbols: Vec<i32> = (0..n).map(|i| i % 100).collect();
         let prices: Vec<f64> = (0..n).map(|i| 50_000.0 + i as f64 * 0.5).collect();
         let volumes: Vec<f64> = (0..n).map(|i| 1.0 + i as f64 * 0.01).collect();
